Move the repository to its own package.
