+++ app/src/main/java/com/goranch/publicapis/ui/food/details/DetailsFoodModule.java
@@ -3,7 +3,7 @@ package com.goranch.publicapis.ui.food.details;
 import com.goranch.publicapis.api.FoodService;
 import com.goranch.publicapis.api.NutritionService;
 import com.goranch.publicapis.di.scopes.FoodDetailScope;
-import com.goranch.publicapis.ui.food.FoodDataRepositoryImpl;
+import com.goranch.publicapis.ui.food.repository.FoodDataRepositoryImpl;
 
 import dagger.Module;
 import dagger.Provides;
