+++ app/src/main/java/com/goranch/publicapis/ui/food/fragment/DetailsFragment.java
@@ -24,11 +24,11 @@ import com.goranch.publicapis.api.ApiComponent;
 import com.goranch.publicapis.api.model.food.nutrition.Food;
 import com.goranch.publicapis.api.model.food.recipe.Recipe;
 import com.goranch.publicapis.di.ComponentProvider;
-import com.goranch.publicapis.ui.food.FoodDataRepositoryImpl;
 import com.goranch.publicapis.ui.food.details.DaggerDetailsFoodComponent;
 import com.goranch.publicapis.ui.food.details.DetailRecipeView;
 import com.goranch.publicapis.ui.food.details.DetailsFoodModule;
 import com.goranch.publicapis.ui.food.details.NutritionRecyclerAdapter;
+import com.goranch.publicapis.ui.food.repository.FoodDataRepositoryImpl;
 import com.goranch.publicapis.ui.food.viewmodel.FoodViewModel;
 import com.goranch.publicapis.ui.util.Utils;
 import com.goranch.publicapis.ui.webview.WebFragment;
